+++ CHANGELOG.md
@@ -11,10 +11,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add ColQwen2.5 support
 - Add support for token pooling with `HierarchicalTokenPooler`.
+- Allow user to specify the maximum number of image tokens in the resized images in `ColQwen2Processor` and `ColQwen2_5_Processor`.
 
 ### Changed
 
 - Warn about evaluation being different from Vidore, and do not store results to prevent confusion.
+- Remove duplicate resize code in `ColQwen2Processor` and `ColQwen2_5_Processor`.
+- Simplify sequence padding for pixel values in `ColQwen2Processor` and `ColQwen2_5_Processor`.
 
 ## [0.3.8] - 2025-01-29